+++ packages/design-tokens/src/utils/color/derivatives.ts
@@ -25,19 +25,29 @@
  */
 
 import { DerivativeColors, SpecifiableColors } from '../../system'
-import { generateInteractive } from './stateful'
-import { mixColors, textBlends } from './blend'
+import { accentBlendScale, generateInteractive } from './stateful'
+import { mixColors, mixScaledColors, textBlends } from './blend'
 
 export const generateDerivativeColors = ({
-  link,
-  text,
   background,
+  inform,
+  link,
+  positive,
+  text,
+  warn,
 }: SpecifiableColors): DerivativeColors => {
+  const accents = {
+    informAccent: mixScaledColors(accentBlendScale, inform, background),
+    positiveAccent: mixScaledColors(accentBlendScale, positive, background),
+    warnAccent: mixScaledColors(accentBlendScale, warn, background),
+  }
+
   return {
     field: background,
     inverse: text,
     inverseOn: background,
     linkInteractive: generateInteractive(link),
     neutral: mixColors(textBlends[1], text, background),
+    ...accents,
   }
 }